revised to fix indexing problem and ensuring that rain timeseries line up
Indexing with an integer does NOT include the right hand index value, so I had to add +1 to the last_tstep_with_rain index when slicing the numpy array rainrate_inperhr. Also, I realized that I was forcing the first timestep with rain for each grid to occur on the first time step rather than preserving the variance in first timestep with rain between the grid cells. Added a section to generate a series of mean rainfall for all overlapping gridcells and use that to identify the first and last indices to use in the time series.